Rename invalid files. Fixed dumb exception.
If a file has an invalid name, the program will create a copy that has a valid name.
Apparently I needed to use a memory stream to fix "A generic error occurred in GDI+" which is a non-descriptive, useless error that has a million different causes and solutions. Thanks for wasting a few hours of my time GDI.